修正resourcetag里的nodeid类型 Signed-off-by: TetsuOu <wangzhe231@huawei.com>

+++ src/gpu/GrGpuResource.h
@@ -116,7 +116,7 @@ struct GrGpuResourceTag {
         isGrGpuResourceTagValid = false;
     }
 
-    GrGpuResourceTag(uint32_t pid, uint32_t tid, uint32_t wid, uint32_t fid, const std::string& name)
+    GrGpuResourceTag(uint32_t pid, uint32_t tid, uint64_t wid, uint32_t fid, const std::string& name)
         : fPid(pid), fTid(tid), fWid(wid), fFid(fid), fName(name)
     {
         isGrGpuResourceTagValid = fPid || fTid || fWid || fFid;
@@ -190,7 +190,7 @@ struct GrGpuResourceTag {
     }
     uint32_t fPid;
     uint32_t fTid;
-    uint32_t fWid;
+    uint64_t fWid;
     uint32_t fFid;
     std::string fName;
     bool isGrGpuResourceTagValid;
